suppress MSVC warning

+++ include/cinolib/io/read_STL.cpp
@@ -120,7 +120,7 @@ void read_STL(const char         * filename,
                     auto it = vmap.find(v);
                     if(it == vmap.end())
                     {
-                        uint fresh_id = vmap.size();
+                        uint fresh_id = uint(vmap.size());
                         vmap[v] = fresh_id;
                         verts.push_back(v);
                         tris.push_back(fresh_id);
@@ -129,7 +129,7 @@ void read_STL(const char         * filename,
                 }
                 else
                 {
-                    tris.push_back(verts.size());
+                    tris.push_back(uint(verts.size()));
                     verts.push_back(v);
                 }
             }
@@ -185,7 +185,7 @@ void read_STL(const char         * filename,
                 }
                 else
                 {
-                    tris.push_back(verts.size());
+                    tris.push_back(uint(verts.size()));
                     verts.push_back(v);
                 }
             }
